refactor: reorganize backend architecture
