+++ apps/backend/src/app.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { ClientModule } from './infra/modules/client/client.module';
+import { ClientModule } from './infra/modules/client.module';
 import { DatabaseModule } from './infra/database/database.module';
 
 @Module({
