+++ apps/backend/src/application/use-cases/client/create-client.use-case.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
-import { Client } from '~/core/entities/client.entity';
-import { ConflictError } from '~/core/errors/conflict.error';
-import { ClientRepository } from '~/core/repositories/client.repository';
+import { Client } from '~/domain/entities/client.entity';
+import { ConflictError } from '~/domain/errors/conflict.error';
+import { ClientRepository } from '~/application/repositories/client.repository';
 
 @Injectable()
 export class CreateClientUseCase {
