+++ apps/backend/src/domain/entities/base.entity.ts
