+++ apps/backend/src/domain/errors/base.error.ts
