+++ apps/backend/src/domain/errors/not-found.error.ts
