+++ apps/backend/src/domain/value-objects/currency.value-object.ts
