+++ apps/backend/src/infra/database/database.module.ts
@@ -1,5 +1,5 @@
 import { Global, Module } from '@nestjs/common';
-import { ClientRepository } from '~/core/repositories/client.repository';
+import { ClientRepository } from '~/application/repositories/client.repository';
 import { InMemoryClientRepository } from '~/test/repositories/in-memory-client.repository';
 
 @Global()
