+++ apps/backend/src/infra/http/mappers/create-client.mapper.ts
@@ -1,7 +1,7 @@
-import { Client } from '~/core/entities/client.entity';
-import { CreateClientDto } from '../dtos/create-client.dto';
-import { Currency } from '~/core/value-objects/currency.value-object';
-import { Id } from '~/core/value-objects/id.value-object';
+import { CreateClientDto } from '~/application/dtos/client/create-client.dto';
+import { Client } from '~/domain/entities/client.entity';
+import { Currency } from '~/domain/value-objects/currency.value-object';
+import { Id } from '~/domain/value-objects/id.value-object';
 
 export class CreateClientMapper {
 	static toDomain(client: CreateClientDto, id?: string): Client {
